requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3
sentence-transformers==2.2.2
torch==2.4.1
faiss-cpu==1.8.0

pandas==2.1.3
numpy==1.26.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
